pyproject.toml
[project]
name = "medusa"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-jalali-date (>=2.0.0,<3.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "medusa", from = "src"}]


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^6.0.0"
django-browser-reload = "^1.18.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
